requirements.txt
alembic==1.0.11
aniso8601==7.0.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
cffi==1.12.3
Click==7.0
colorama==0.4.1
cryptography==2.7
entrypoints==0.3
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.4.1
Mako==1.0.12
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.3.1
pylint-plugin-utils==0.5
pyrsistent==0.15.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
